+++ src/Make_morph.mak
@@ -102,6 +102,7 @@ SRC =	arabic.c						\
 	term.c							\
 	testing.c						\
 	textprop.c						\
+	time.c							\
 	ui.c							\
 	undo.c							\
 	usercmd.c						\
