+++ src/Make_vms.mms
@@ -381,6 +381,7 @@ SRC = \
 	termlib.c \
 	testing.c \
 	textprop.c \
+	time.c \
 	ui.c \
 	undo.c \
 	usercmd.c \
@@ -487,6 +488,7 @@ OBJ = \
 	termlib.obj \
 	testing.obj \
 	textprop.obj \
+	time.obj \
 	ui.obj \
 	undo.obj \
 	usercmd.obj \
@@ -976,6 +978,9 @@ testing.obj : testing.c vim.h [.auto]config.h feature.h os_unix.h   \
  ascii.h keymap.h term.h macros.h structs.h regexp.h gui.h beval.h \
  [.proto]gui_beval.pro option.h ex_cmds.h proto.h globals.h
 textprop.obj : textprop.c vim.h [.auto]config.h feature.h os_unix.h   \
+ ascii.h keymap.h term.h macros.h structs.h regexp.h gui.h beval.h \
+ [.proto]gui_beval.pro option.h ex_cmds.h proto.h globals.h
+time.obj : time.c vim.h [.auto]config.h feature.h os_unix.h   \
  ascii.h keymap.h term.h macros.h structs.h regexp.h gui.h beval.h \
  [.proto]gui_beval.pro option.h ex_cmds.h proto.h globals.h
 ui.obj : ui.c vim.h [.auto]config.h feature.h os_unix.h   \
