+++ src/Makefile
@@ -1661,6 +1661,7 @@ BASIC_SRC = \
 	terminal.c \
 	testing.c \
 	textprop.c \
+	time.c \
 	ui.c \
 	undo.c \
 	usercmd.c \
@@ -1800,6 +1801,7 @@ OBJ_COMMON = \
 	objects/terminal.o \
 	objects/testing.o \
 	objects/textprop.o \
+	objects/time.o \
 	objects/ui.o \
 	objects/undo.o \
 	objects/usercmd.o \
@@ -1971,6 +1973,7 @@ PRO_AUTO = \
 	termlib.pro \
 	testing.pro \
 	textprop.pro \
+	time.pro \
 	ui.pro \
 	undo.pro \
 	usercmd.pro \
@@ -3433,6 +3436,9 @@ objects/testing.o: testing.c
 objects/textprop.o: textprop.c
 	$(CCC) -o $@ textprop.c
 
+objects/time.o: time.c
+	$(CCC) -o $@ time.c
+
 objects/ui.o: ui.c
 	$(CCC) -o $@ ui.c
 
@@ -3956,6 +3962,10 @@ objects/testing.o: testing.c vim.h protodef.h auto/config.h feature.h os_unix.h 
  proto/gui_beval.pro structs.h regexp.h gui.h alloc.h ex_cmds.h spell.h \
  proto.h globals.h
 objects/textprop.o: textprop.c vim.h protodef.h auto/config.h feature.h os_unix.h \
+ auto/osdef.h ascii.h keymap.h term.h macros.h option.h beval.h \
+ proto/gui_beval.pro structs.h regexp.h gui.h alloc.h ex_cmds.h spell.h \
+ proto.h globals.h
+objects/time.o: time.c vim.h protodef.h auto/config.h feature.h os_unix.h \
  auto/osdef.h ascii.h keymap.h term.h macros.h option.h beval.h \
  proto/gui_beval.pro structs.h regexp.h gui.h alloc.h ex_cmds.h spell.h \
  proto.h globals.h
