+++ src/proto.h
@@ -221,6 +221,7 @@ void mbyte_im_set_active(int active_arg);
 #  include "textprop.pro"
 # endif
 # include "testing.pro"
+# include "time.pro"
 # include "ui.pro"
 # include "undo.pro"
 # include "usercmd.pro"
