+++ src/proto/ex_cmds2.pro
@@ -1,15 +1,4 @@
 /* ex_cmds2.c */
-long proftime_time_left(proftime_T *due, proftime_T *now);
-timer_T *create_timer(long msec, int repeat);
-long check_due_timer(void);
-void stop_timer(timer_T *timer);
-int set_ref_in_timer(int copyID);
-void timer_free_all(void);
-void f_timer_info(typval_T *argvars, typval_T *rettv);
-void f_timer_pause(typval_T *argvars, typval_T *rettv);
-void f_timer_start(typval_T *argvars, typval_T *rettv);
-void f_timer_stop(typval_T *argvars, typval_T *rettv);
-void f_timer_stopall(typval_T *argvars, typval_T *rettv);
 int autowrite(buf_T *buf, int forceit);
 void autowrite_all(void);
 int check_changed(buf_T *buf, int flags);
